WIP: [ios] Moving map downloader to SwiftUI

+++ iphone/CoreApi/CoreApi/Storage/MWMMapNodeAttributes.h
@@ -33,7 +33,7 @@ NS_SWIFT_NAME(MapNodeAttributes)
 @property(nonatomic, readonly) NSUInteger downloadingMwmCount;
 @property(nonatomic, readonly) uint64_t totalSize;
 @property(nonatomic, readonly) uint64_t downloadedSize;
-@property(nonatomic, readonly) uint64_t downloadingSize;
+@property(nonatomic, readonly) double downloadingSize;
 @property(nonatomic, readonly) uint64_t totalUpdateSizeBytes;
 @property(nonatomic, readonly) NSString *nodeName;
 @property(nonatomic, readonly) NSString *nodeDescription;
